fix(workspace): dedupe members; refine ZK eligibility verifier interface
- Remove duplicate care-plan/pacs-integration workspace entries (Cargo fingerprint issues)
- Split trait/placeholder into interface.rs; add PublicInputs alias and RUST_INTERFACE_VERSION
- Add verify_eligibility_proof free function; document labels in crate rustdoc
- Add INTERFACE.md for proof-system boundary, actors, and future tests
- Extend integration tests (version check, free function parity)
- zk-eligibility-verifier: use rlib only to avoid duplicate lib artifact warning

Made-with: Cursor
