+++ Cargo.toml
@@ -25,11 +25,7 @@ members = [
   "contracts/care-plan",
   "contracts/pacs-integration",
   "contracts/healthcare-credentialing",
-  "contracts/care-plan" ,
-  "contracts/pacs-integration",
   "contracts/nutrition-care-management",
-  "contracts/care-plan",
-  "contracts/pacs-integration",
   "contracts/dental-records",
   "contracts/mental-health",
   "contracts/rehabilitation-services",
